mobile/android/tests/browser/chrome/test_session_scroll_position.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=810981
https://bugzilla.mozilla.org/show_bug.cgi?id=1282902
https://bugzilla.mozilla.org/show_bug.cgi?id=1301016
https://bugzilla.mozilla.org/show_bug.cgi?id=1265818
-->
<head>
  <meta charset="utf-8">
  <title>Tests for Bug 810981, 1282902, 1301016, 1265818</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript">

  /** Tests for Bug 810981, 1282902, 1301016, 1265818 **/

  "use strict";

  ChromeUtils.import("resource://gre/modules/XPCOMUtils.jsm");
  ChromeUtils.import("resource://gre/modules/Services.jsm");
  ChromeUtils.import("resource://gre/modules/Messaging.jsm");
  ChromeUtils.import("resource://gre/modules/Task.jsm");

  // The chrome window and friends.
  let chromeWin = Services.wm.getMostRecentWindow("navigator:browser");
  let BrowserApp = chromeWin.BrowserApp;

  // Use something with ample space for scrolling and zooming.
  const URL = "http://example.org/chrome/mobile/android/tests/browser/chrome/browser_scrollPositions_sample.html";
  // Same content but under a different URL, so we have something to navigate to and from.
  const URL2 = "http://example.org/chrome/mobile/android/tests/browser/chrome/browser_scrollPositions_sample2.html";
  // Reader mode URL
  const URL_reader = "about:reader?url=http%3A%2F%2Fexample.org%2Fchrome%2Fmobile%2Fandroid%2Ftests%2Fbrowser%2Fchrome%2Fbasic_article_mobile.html";

  // Randomized set of scroll positions and zoom factors we will use in this test.
  const SCROLL_X = Math.round(100 * (1 + Math.random()));
  const SCROLL_Y = Math.round(200 * (1 + Math.random()));
  const SCROLL_STR = SCROLL_X + "," + SCROLL_Y;
  const SCROLL_STR_Y_ONLY = 0 + "," + SCROLL_Y;
  const ZOOM = 1 + 0.5 * Math.random();

  const SCROLL2_X = Math.round(300 * (1 + Math.random()));
  const SCROLL2_Y = Math.round(400 * (1 + Math.random()));
  const SCROLL2_STR = SCROLL2_X + "," + SCROLL2_Y;
  const ZOOM2 = 1.5 + 0.5 * Math.random();

  function dispatchUIEvent(browser, type) {
    let event = browser.contentDocument.createEvent("UIEvents");
    event.initUIEvent(type, true, false, browser.contentDocument.defaultView, 0);
    browser.dispatchEvent(event);
  }

  function setScrollPosition(browser, x, y) {
    browser.contentWindow.scrollTo(x, y);
    dispatchUIEvent(browser, "scroll");
  }

  function setZoomLevel(browser, zoom) {
    browser.contentWindow.QueryInterface(
      Ci.nsIInterfaceRequestor).getInterface(
        Ci.nsIDOMWindowUtils).setResolutionAndScaleTo(zoom);
  }

  // Track the tabs where the tests are happening.
  let tabScroll;

  function cleanupTabs() {
    if (tabScroll) {
      BrowserApp.closeTab(tabScroll);
      tabScroll = null;
    }
  }

  SimpleTest.registerCleanupFunction(function() {
    cleanupTabs();
  });

  let ss = Cc["@mozilla.org/browser/sessionstore;1"].getService(Ci.nsISessionStore);

  add_task(async function test_sessionStoreScrollPositionReaderMode() {
    // Creates a tab, sets a scroll position and closes the tab.
    function createAndRemoveReaderTab() {
      return Task.spawn(function* () {
        // Create a new tab.
        tabScroll = BrowserApp.addTab(URL_reader);
        let browser = tabScroll.browser;
        yield promiseBrowserEvent(browser, "AboutReaderContentReady");

        // Modify scroll position.
        setScrollPosition(browser, 0, SCROLL_Y);
        yield promiseTabEvent(browser, "SSTabScrollCaptured");

        // Check that we've actually scrolled.
        let ifreq = browser.contentWindow.QueryInterface(Ci.nsIInterfaceRequestor);
        let utils = ifreq.getInterface(Ci.nsIDOMWindowUtils);
        let scrollX = {}, scrollY = {};
        utils.getScrollXY(false, scrollX, scrollY);
        is(scrollX.value, 0, "scrollX set correctly");
        is(scrollY.value, SCROLL_Y, "scrollY set correctly");

        // Remove the tab.
        BrowserApp.closeTab(tabScroll);
        yield promiseTabEvent(browser, "SSTabCloseProcessed");
      });
    }

    await createAndRemoveReaderTab();
    let state = ss.getClosedTabs(chromeWin);
    let [{scrolldata}] = state;
    is(scrolldata.scroll, SCROLL_STR_Y_ONLY, "stored scroll position is correct");

    // Restore the closed tab.
    let closedTabData = ss.getClosedTabs(chromeWin)[0];
    let browser = ss.undoCloseTab(chromeWin, closedTabData);
    await promiseBrowserEvent(browser, "AboutReaderContentReady");

    // Check the scroll position.
    let ifreq = browser.contentWindow.QueryInterface(Ci.nsIInterfaceRequestor);
    let utils = ifreq.getInterface(Ci.nsIDOMWindowUtils);
    let scrollX = {}, scrollY = {};
    utils.getScrollXY(false, scrollX, scrollY);
    is(scrollX.value, 0, "scrollX restored correctly");
    is(scrollY.value, SCROLL_Y, "scrollY restored correctly");

    // Remove the tab.
    BrowserApp.closeTab(BrowserApp.getTabForBrowser(browser));
  });

  add_task(async function test_sessionStoreScrollPositionAndZoomLevel() {
    // Creates a tab, sets a scroll position and zoom level and closes the tab.
    function createAndRemoveTab() {
      return Task.spawn(function* () {
        // Create a new tab.
        tabScroll = BrowserApp.addTab(URL);
        let browser = tabScroll.browser;
        yield promiseBrowserEvent(browser, "pageshow");

        // Modify scroll position and zoom level.
        setZoomLevel(browser, ZOOM);
        setScrollPosition(browser, SCROLL_X, SCROLL_Y);
        yield promiseTabEvent(browser, "SSTabScrollCaptured");

        // Check that we've actually scrolled and zoomed.
        let ifreq = browser.contentWindow.QueryInterface(Ci.nsIInterfaceRequestor);
        let utils = ifreq.getInterface(Ci.nsIDOMWindowUtils);
        let scrollX = {}, scrollY = {}, zoom = {};
        utils.getResolution(zoom);
        utils.getScrollXY(false, scrollX, scrollY);
        ok(fuzzyEquals(zoom.value, ZOOM), "zoom set correctly");
        is(scrollX.value, SCROLL_X, "scrollX set correctly");
        is(scrollY.value, SCROLL_Y, "scrollY set correctly");

        // Navigate to a different page and scroll/zoom there as well.
        browser.loadURI(URL2);
        yield promiseBrowserEvent(browser, "pageshow");

        setZoomLevel(browser, ZOOM2);
        setScrollPosition(browser, SCROLL2_X, SCROLL2_Y);
        yield promiseTabEvent(browser, "SSTabScrollCaptured");

        utils.getResolution(zoom);
        utils.getScrollXY(false, scrollX, scrollY);
        ok(fuzzyEquals(zoom.value, ZOOM2), "zoom set correctly");
        is(scrollX.value, SCROLL2_X, "scrollX set correctly");
        is(scrollY.value, SCROLL2_Y, "scrollY set correctly");

        // Remove the tab.
        BrowserApp.closeTab(tabScroll);
        yield promiseTabEvent(browser, "SSTabCloseProcessed");
      });
    }

    await createAndRemoveTab();
    let state = ss.getClosedTabs(chromeWin);
    let [{scrolldata}] = state;
    is(scrolldata.scroll, SCROLL2_STR, "stored scroll position is correct");
    ok(fuzzyEquals(scrolldata.zoom.resolution, ZOOM2), "stored zoom level is correct");

    // Restore the closed tab.
    let closedTabData = ss.getClosedTabs(chromeWin)[0];
    let browser = ss.undoCloseTab(chromeWin, closedTabData);
    let pageshow = promiseBrowserEvent(browser, "pageshow");
    let scroll = promiseBrowserEvent(browser, "scroll");
    await pageshow;
    await scroll;

    // Check the scroll position and zoom level.
    let ifreq = browser.contentWindow.QueryInterface(Ci.nsIInterfaceRequestor);
    let utils = ifreq.getInterface(Ci.nsIDOMWindowUtils);
    let scrollX = {}, scrollY = {}, zoom = {};
    utils.getResolution(zoom);
    utils.getScrollXY(false, scrollX, scrollY);
    ok(fuzzyEquals(zoom.value, ZOOM2), "zoom restored correctly");
    is(scrollX.value, SCROLL2_X, "scrollX restored correctly");
    is(scrollY.value, SCROLL2_Y, "scrollY restored correctly");

    // Now go back in history and check that the scroll position
    // is restored there as well.
    is(browser.canGoBack, true, "can go back");
    pageshow = promiseBrowserEvent(browser, "pageshow");
    scroll = promiseBrowserEvent(browser, "scroll");
    browser.goBack();
    await pageshow;
    await scroll;

    utils.getResolution(zoom);
    utils.getScrollXY(false, scrollX, scrollY);
    ok(fuzzyEquals(zoom.value, ZOOM), "zoom restored correctly");
    is(scrollX.value, SCROLL_X, "scrollX restored correctly");
    is(scrollY.value, SCROLL_Y, "scrollY restored correctly");

    // Remove the tab.
    BrowserApp.closeTab(BrowserApp.getTabForBrowser(browser));
  });

  add_task(async function test_sessionStoreZoomLevelRecalc() {
    // Creates a tab, sets a scroll position and zoom level and closes the tab.
    function createAndRemoveTab() {
      return Task.spawn(function* () {
        // Create a new tab.
        tabScroll = BrowserApp.addTab(URL);
        let browser = tabScroll.browser;
        yield promiseBrowserEvent(browser, "pageshow");

        // Modify scroll position and zoom level.
        setZoomLevel(browser, ZOOM);
        setScrollPosition(browser, SCROLL_X, SCROLL_Y);
        yield promiseTabEvent(browser, "SSTabScrollCaptured");

        // Check that we've actually scrolled and zoomed.
        let ifreq = browser.contentWindow.QueryInterface(Ci.nsIInterfaceRequestor);
        let utils = ifreq.getInterface(Ci.nsIDOMWindowUtils);
        let scrollX = {}, scrollY = {}, zoom = {};
        utils.getResolution(zoom);
        utils.getScrollXY(false, scrollX, scrollY);
        ok(fuzzyEquals(zoom.value, ZOOM), "zoom set correctly");
        is(scrollX.value, SCROLL_X, "scrollX set correctly");
        is(scrollY.value, SCROLL_Y, "scrollY set correctly");

        // Remove the tab.
        BrowserApp.closeTab(tabScroll);
        yield promiseTabEvent(browser, "SSTabCloseProcessed");
      });
    }

    await createAndRemoveTab();
    let state = ss.getClosedTabs(chromeWin);
    let [{scrolldata}] = state;
    is(scrolldata.scroll, SCROLL_STR, "stored scroll position is correct");
    ok(fuzzyEquals(scrolldata.zoom.resolution, ZOOM), "stored zoom level is correct");

    // Pretend the zoom level was originally saved on a rotated device.
    let closedTabData = ss.getClosedTabs(chromeWin)[0];
    let displayWidth = scrolldata.zoom.displaySize.width;
    let displayHeight = scrolldata.zoom.displaySize.height;
    closedTabData.scrolldata.zoom.displaySize.width = displayHeight;
    closedTabData.scrolldata.zoom.displaySize.height = displayWidth;

    // Restore the closed tab.
    let browser = ss.undoCloseTab(chromeWin, closedTabData);
    await promiseBrowserEvent(browser, "pageshow");

    // Check the scroll position and zoom level.
    let ifreq = browser.contentWindow.QueryInterface(Ci.nsIInterfaceRequestor);
    let utils = ifreq.getInterface(Ci.nsIDOMWindowUtils);
    let scrollX = {}, scrollY = {}, zoom = {};
    utils.getResolution(zoom);
    utils.getScrollXY(false, scrollX, scrollY);
    ok(fuzzyEquals(zoom.value, ZOOM * displayWidth / displayHeight), "recalculated zoom restored correctly");
    is(scrollX.value, SCROLL_X, "scrollX restored correctly");
    is(scrollY.value, SCROLL_Y, "scrollY restored correctly");

    // Remove the tab.
    BrowserApp.closeTab(BrowserApp.getTabForBrowser(browser));
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=810981">Mozilla Bug 810981</a>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1282902">Mozilla Bug 1282902</a>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1301016">Mozilla Bug 1301016</a>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1265818">Mozilla Bug 1265818</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
